Move 00-begin to src/workshop
