--- 00-begin/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "react-workshop",
-  "version": "0.1.0",
-  "private": true,
-  "engines": {
-    "node": ">=6"
-  },
-  "dependencies": {
-    "body-parser": "^1.17.2",
-    "classnames": "^2.2.5",
-    "express": "^4.15.4",
-    "lodash": "^4.17.4",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.6",
-    "react-scripts": "1.1.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "start:api": "node api-server.js"
-  }
-}
