+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './end/App';
+import App from './workshop/App';
 
 import './index.css';
