+++ src/workshop/components/EmailView.css
